PrefUI // Boost SwiftUI requirements to macOS 13.
