+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneBehavior.swift
@@ -12,7 +12,7 @@ import SwiftExtension
 import SwiftUI
 import SwiftUIBackports
 
-@available(macOS 10.15, *)
+@available(macOS 13, *)
 struct VwrPrefPaneBehavior: View {
   // MARK: - AppStorage Variables
 
@@ -217,7 +217,7 @@ struct VwrPrefPaneBehavior: View {
             )
             Spacer()
             Text(
-              "This feature requires macOS 10.15 and above.".localized + CtlPrefUIShared.sentenceSeparator
+              "This feature requires macOS 10.15 and above.".localized + CtlPrefUI.sentenceSeparator
                 + "This feature only needs to parse consecutive NSEvents passed by macOS built-in InputMethodKit framework, hence no necessity of asking end-users for extra privileges of monitoring global keyboard inputs. You are free to investigate our codebase or reverse-engineer this input method to see whether the above statement is trustable.".localized
             )
             .settingsDescription()
@@ -251,13 +251,13 @@ struct VwrPrefPaneBehavior: View {
             .settingsDescription()
           }
         }
-      }.formStyled().frame(minWidth: CtlPrefUIShared.formWidth, maxWidth: ceil(CtlPrefUIShared.formWidth * 1.2))
+      }.formStyled().frame(minWidth: CtlPrefUI.formWidth, maxWidth: ceil(CtlPrefUI.formWidth * 1.2))
     }
-    .frame(maxHeight: CtlPrefUIShared.contentMaxHeight)
+    .frame(maxHeight: CtlPrefUI.contentMaxHeight)
   }
 }
 
-@available(macOS 11.0, *)
+@available(macOS 13, *)
 struct VwrPrefPaneBehavior_Previews: PreviewProvider {
   static var previews: some View {
     VwrPrefPaneBehavior()
