+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneCandidates.swift
@@ -12,7 +12,7 @@ import SwiftExtension
 import SwiftUI
 import SwiftUIBackports
 
-@available(macOS 10.15, *)
+@available(macOS 13, *)
 struct VwrPrefPaneCandidates: View {
   // MARK: - AppStorage Variables
 
@@ -190,13 +190,13 @@ struct VwrPrefPaneCandidates: View {
             isOn: $enableMouseScrollingForTDKCandidatesCocoa
           )
         }
-      }.formStyled().frame(minWidth: CtlPrefUIShared.formWidth, maxWidth: ceil(CtlPrefUIShared.formWidth * 1.2))
+      }.formStyled().frame(minWidth: CtlPrefUI.formWidth, maxWidth: ceil(CtlPrefUI.formWidth * 1.2))
     }
-    .frame(maxHeight: CtlPrefUIShared.contentMaxHeight)
+    .frame(maxHeight: CtlPrefUI.contentMaxHeight)
   }
 }
 
-@available(macOS 11.0, *)
+@available(macOS 13, *)
 struct VwrPrefPaneCandidates_Previews: PreviewProvider {
   static var previews: some View {
     VwrPrefPaneCandidates()
@@ -205,7 +205,7 @@ struct VwrPrefPaneCandidates_Previews: PreviewProvider {
 
 // MARK: - Selection Key Preferences (View)
 
-@available(macOS 10.15, *)
+@available(macOS 13, *)
 private struct VwrPrefPaneCandidates_SelectionKeys: View {
   // MARK: - AppStorage Variables
