+++ Source/Modules/UIModules/PrefUI/VwrPrefPaneDictionary.swift
@@ -14,7 +14,7 @@ import SwiftExtension
 import SwiftUI
 import SwiftUIBackports
 
-@available(macOS 10.15, *)
+@available(macOS 13, *)
 struct VwrPrefPaneDictionary: View {
   // MARK: - AppStorage Variables
 
@@ -232,13 +232,13 @@ struct VwrPrefPaneDictionary: View {
             .settingsDescription()
           }
         }
-      }.formStyled().frame(minWidth: CtlPrefUIShared.formWidth, maxWidth: ceil(CtlPrefUIShared.formWidth * 1.2))
+      }.formStyled().frame(minWidth: CtlPrefUI.formWidth, maxWidth: ceil(CtlPrefUI.formWidth * 1.2))
     }
-    .frame(maxHeight: CtlPrefUIShared.contentMaxHeight)
+    .frame(maxHeight: CtlPrefUI.contentMaxHeight)
   }
 }
 
-@available(macOS 11.0, *)
+@available(macOS 13, *)
 struct VwrPrefPaneDictionary_Previews: PreviewProvider {
   static var previews: some View {
     VwrPrefPaneDictionary()
