+++ Source/Modules/UIModules/PrefUI/VwrPrefPanePhrases.swift
@@ -12,7 +12,7 @@ import Shared
 import SwiftExtension
 import SwiftUI
 
-@available(macOS 10.15, *)
+@available(macOS 13, *)
 struct VwrPrefPanePhrases: View {
   var body: some View {
     ScrollView {
@@ -26,13 +26,13 @@ struct VwrPrefPanePhrases: View {
       }
       .padding(4)
       .padding()
-      .frame(minWidth: CtlPrefUIShared.formWidth, maxWidth: ceil(CtlPrefUIShared.formWidth * 1.2))
+      .frame(minWidth: CtlPrefUI.formWidth, maxWidth: ceil(CtlPrefUI.formWidth * 1.2))
     }
-    .frame(maxHeight: CtlPrefUIShared.contentMaxHeight)
+    .frame(maxHeight: CtlPrefUI.contentMaxHeight)
   }
 }
 
-@available(macOS 11.0, *)
+@available(macOS 13, *)
 struct VwrPrefPanePhrases_Previews: PreviewProvider {
   static var previews: some View {
     VwrPrefPanePhrases()
